# swagger.yaml
openapi: 3.1.0
info:
  title: T-shirt Store Intenship Project
  description: Creating an API for e-commerce
  version: 1.1.0
  contact: 
    email: ariffidan03@gmail.com
    url: "https://scramblebit.com/"

# servers:
#   - url: "https://localhost:4000/api/v1"
#     description: for local host - secure
#   - url: "http://localhost:4000/api/v1"
#     description: for local host - regular
servers:
  - url: "{protocol}://localhost:4000/api/{version}"
    description: for local host - secure
    variables: 
      version:
        enum:
          - v1
          - v2
        default: v1
      protocol:
        enum:
          - http
          - https
        default: http
components: 
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /admin/products/{id}:
    put:
      summary: Update a specific product by admin
      tags:
        - Product
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the product to update
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Product name
                  example: "New Product"
                description:
                  type: string
                  description: Product description
                  example: "Updated description for the product"
                price:
                  type: number
                  format: float
                  description: Product price
                  example: 19.99
                photos:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: string
                      secure_url:
                        type: string
                  description: Array of product photo objects
      responses:
        '200':
          description: Product successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  product:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      description:
                        type: string
                      price:
                        type: number
                        format: float
                      photos:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            secure_url:
                              type: string
        '404':
          description: Product not found
        '500':
          description: Internal server error
    delete:
      summary: Delete a specific product by admin
      operationId: adminDeleteSingleProduct
      tags:
        - Product
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the product to delete
          schema:
            type: string
      responses:
        '200':
          description: Product successfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "The product ProductName was deleted successfully!"
        '404':
          description: Product not found
        '500':
          description: Internal server error

  # order part
  /order/create:
    post:
      summary: Create a new order
      operationId: createOrder
      tags:
        - Order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shippingInfo:
                  type: object
                  description: Shipping information
                  properties:
                    address:
                      type: string
                    city:
                      type: string
                    postalCode:
                      type: string
                    country:
                      type: string
                orderItems:
                  type: array
                  items:
                    type: object
                    properties:
                      productId:
                        type: string
                      quantity:
                        type: integer
                paymentInfo:
                  type: object
                  description: Payment information
                  properties:
                    paymentMethod:
                      type: string
                      example: "Credit Card"
                taxAmount:
                  type: number
                  format: float
                  description: Tax amount for the order
                shippingAmount:
                  type: number
                  format: float
                  description: Shipping fee for the order
                totalAmount:
                  type: number
                  format: float
                  description: Total amount of the order
      responses:
        '200':
          description: Order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  order:
                    type: object
                    properties:
                      _id:
                        type: string
                      shippingInfo:
                        type: object
                      orderItems:
                        type: array
                      paymentInfo:
                        type: object
                      taxAmount:
                        type: number
                      shippingAmount:
                        type: number
                      totalAmount:
                        type: number
        '400':
          description: Invalid request data
        '500':
          description: Internal server error
...